Start implement PNG
In a previous nodejs project, I found there is noe PNG in node js, so
I need to implement one. From a list of PNG algorithm on wiki, I found
ACORN is the simplist one to understand and implement.

It's very simple as a few lines of math:
https://acorn.wikramaratna.org/

From the result I got from that project, it's good enough. But I've
never tested if it's REALLY random or not. It's a good chance to
implement it in C again and test it, with Monte Carlo or something else